open-spec-ai-news-feed/backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
